requirements.txt
requests>=2.31.0
flask>=3.0.0
flask-login>=0.6.3
caldav>=1.3.9
icalendar>=5.0.11
python-dateutil>=2.8.2
lxml>=4.9.3
playwright>=1.40.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
